# gateway/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "US Government Data API",
    "description": "Unified REST API providing access to 45 US government data endpoints across 13 categories including cybersecurity vulnerability intelligence. Real-time data from NVD, CISA KEV, EPSS, NHTSA, FDA, CFPB, FEMA, NOAA, USGS, SEC, BLS, USDA, EPA, HUD, FAA, CBP, SAM.gov, CourtListener, CDC, WHO, Federal Register, Grants.gov, OpenCorporates, and more.",
    "version": "1.1.0",
    "contact": {
      "name": "MARTC03",
      "url": "https://apify.com/martc03"
    }
  },
  "servers": [
    {
      "url": "https://govdata-api.netlify.app",
      "description": "Production (Netlify Proxy)"
    }
  ],
  "paths": {
...
    "/api/v1/environment/hud-foreclosures": {
      "get": {
        "operationId": "searchHudForeclosures",
        "summary": "Search HUD Foreclosures",
        "description": "Search HUD foreclosure property listings by state and county.",
        "tags": ["Environment"],
        "parameters": [
          { "name": "state", "in": "query", "schema": { "type": "string" }, "description": "State" },
          { "name": "county", "in": "query", "schema": { "type": "string" }, "description": "County" },
          { "name": "limit", "in": "query", "schema": { "type": "integer" }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
...
  },
  "components": {
    "responses": {
      "SuccessResponse": {
        "description": "Successful response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": true },
                "data": { "type": "object" },
                "meta": {
                  "type": "object",
                  "properties": {
                    "tool": { "type": "string" },
                    "latencyMs": { "type": "integer" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string" },
                "meta": {
                  "type": "object",
                  "properties": {
                    "tool": { "type": "string" },
                    "latencyMs": { "type": "integer" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "RapidAPIKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key"
      }
    }
  },
  "security": [
    { "RapidAPIKey": [] }
  ],
...
}